requirements.txt
Flask==2.3.3
gunicorn==23.0.0
requests
beautifulsoup4
yt-dlp
chat-downloader
scrapetube
python-dotenv
